requirements.txt
flask==2.1.0
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
requests==2.26.0
python-dotenv==0.19.2
pytest==6.2.5
gunicorn==20.1.0